fonts: speculative fix for Monaco.dfont bitmap/render error
The error message in this issue sounds a lot like the freetype
regression that caused bitmap fonts to fail to render in a recent
freetype release.

Our workaround for that is to used our understanding of whether
a font is a bitmap font or not to avoid calling render.

What we normally see for bitmap TTFs is that setting the scale
fails and we then fall back to using a bitmap.

For Monaco.dfont it appears as though setting the scale succeeds.

This commit introduces some skepticism and prefers to use bitmaps
when available.

This might potentially cause problems in the future if there are
fonts that legitimately have both scaled and bitmap fonts, but
lets see if this helps for now.

refs: https://github.com/wez/wezterm/issues/1419