font: improve alpha channel in rasterizer
Previously, we'd set it to 100% if any non-zero channel was present,
but this resulted in stronger/darker outlines that were especially
noticeable when the window background was transparent and over a
bright background.

This commit sets the alpha based on the coverage provided to us
by freetype and looks better with and without a transparent window
background.

@dmfay: I think this might be behind what you noticed in your last
comment on #1025.

refs: #1325