mux: don't include unchanged lines in the bonus lines
This doesn't make a huge difference; I think we already addressed
the worst of the overhead with a combination of
https://github.com/wez/wezterm/commit/eae327efccc5e8de25c1fdafe438f802af185afe
and the more recent commits to avoid over invalidation around
cursor movement and screen scrolling.

The numbers I see are a very small number of bytes per second when
scrolling through eg: vim and aren't crazy or alarming.

So I'm going to consider the related issue resolved!

closes: https://github.com/wez/wezterm/issues/867