ci: migrate centos:8 images to stream
The centos8 builds have been failing today, and it appears to
be because https://www.centos.org/centos-linux-eol/ happened
about 3 years early and the URLs in the image are hard-broken.

https://forums.centos.org/viewtopic.php?f=54&t=78708 has some
info on re-pointing them to a vault subdomain.

The "right" thing to do is to migrate to centos 8 stream, but
I don't see a centos stream docker image from centos.

This commit attempts the manual migration steps to stream.

Let's see how this goes.