shaping: improve glyph cell width math
This is a more robust approach; we make a separate pass to figure
out information about the (harfbuzz) cluster for a sequence of glyphs,
and then map that sequence back to the original cell sequence, and
from there compute the total cell width for the run, then distribute
the glyphs across the run.

This should yield more sane results for bidi.

Fixup the x-position math; it was still wonky despite the
efforts in 5f2c905db86353040c63ee9335cf69ca9fd5e99c and
af92265ffb214878beadb543bd9e4b74e0bb8a99

refs: #1570
refs: #1607
refs: #1563