Open directories in existing terminal windows
Instead of opening every directory in a new wezterm instance check if
there's a running wezterm instance and if so, spawn every directory into
a new tab in that instance.  Otherwise spawn a new instance for the
first directory, and then spawn the remaining directories into that
instance.

Also extract the systemd interface from the menu provider class, to make
things easier to understand.