fonts: workaround broken symbol fonts with 0 advances
The broot icon font has glyphs with horizontal advance set to 0.  That
would cause us to consider the glyph to be zero width, so handle that as
a special case.  Note that it is legit for certain cells to end up with
a zero advance/width during shaping if they represent combining
characters: this is more common in Arabic scripts.

refs: https://github.com/wez/wezterm/issues/1787