macos: more urgh with swapped chars/unmod
a redo of 81d08325c141ced838271e9f3d28c8271ee4cc5f, which was a bit
more nuanced. Partially revert the logic from that commit and rephrase
it as factoring out the condition to make that part clearer, but restore
the original action, which is important in order to propagate the
modifiers through.

refs: https://github.com/wez/wezterm/issues/1907#issuecomment-1108797343